extract some config entry helpers into package
